+++ src/main/java/org/folio/rest/impl/PatronServicesResourceImpl.java
@@ -12,6 +12,7 @@ import org.folio.patron.rest.exceptions.ModuleGeneratedHttpException;
 import org.folio.rest.RestVerticle;
 import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.model.Account;
+import org.folio.rest.jaxrs.model.Charge;
 import org.folio.rest.jaxrs.model.Hold;
 import org.folio.rest.jaxrs.model.Hold.FulfillmentPreference;
 import org.folio.rest.jaxrs.model.Hold.Status;
@@ -53,10 +54,6 @@ public class PatronServicesResourceImpl implements PatronServicesResource {
 
             account.setTotalCharges(new TotalCharges().withAmount(0.0).withIsoCurrencyCode("USD"));
             account.setTotalChargesCount(0);
-            // TODO: When mod-feesfines is available, do something here
-//            if (includeCharges) {
-//              // call feesfines here
-//            }
 
             final CompletableFuture<Account> cf1 = httpClient.request("/circulation/loans?limit=" + getLimit(includeLoans) + "&query=%28userId%3D%3D" + id + "%20and%20status.name%3D%3DOpen%29", okapiHeaders)
                 .thenApply(response -> {
@@ -70,7 +67,24 @@ public class PatronServicesResourceImpl implements PatronServicesResource {
                   return addHolds(account, response, includeHolds);
                 });
 
-            return CompletableFuture.allOf(cf1, cf2)
+            final CompletableFuture<Account> cf3 = httpClient.request("/accounts?limit=" + getLimit(true) + "&query=%28userId%3D%3D" + id + "%20and%20status.name%3D%3DOpen%29", okapiHeaders)
+                .thenApply(response -> {
+                  verifyExists(response);
+                  return addCharges(account, response, includeCharges);
+                }).thenCompose(charges -> {
+                  if (includeCharges) {
+                    List<CompletableFuture<Account>> cfs = new ArrayList<>();
+                    for (Charge charge: account.getCharges()) {
+                      cfs.add(lookupChargeItem(httpClient, charge, account, okapiHeaders));
+                      cfs.add(lookupFeeFine(httpClient, charge, account, okapiHeaders));
+                    }
+                    return CompletableFuture.allOf(cfs.toArray(new CompletableFuture[cfs.size()]))
+                        .thenApply(done -> account);
+                  }
+                  return CompletableFuture.completedFuture(account);
+                });
+
+            return CompletableFuture.allOf(cf1, cf2, cf3)
                 .thenApply(result -> account);
           } catch (Exception e) {
             throw new CompletionException(e);
@@ -317,6 +331,158 @@ public class PatronServicesResourceImpl implements PatronServicesResource {
         .withRequestId(holdJson.getString("id"))
         .withFulfillmentPreference(FulfillmentPreference.fromValue(holdJson.getString("fulfilmentPreference")))
         .withStatus(Status.fromValue(holdJson.getString("status")));
+  }
+
+  private Account addCharges(Account account, Response response, boolean includeCharges) {
+    final int totalCharges = response.getBody().getInteger("totalRecords", Integer.valueOf(0)).intValue();
+    final List<Charge> charges = new ArrayList<>();
+
+    account.setTotalChargesCount(totalCharges);
+    account.setCharges(charges);
+
+    double amount = 0.0;
+
+    if (totalCharges > 0) {
+      final JsonArray accountsJson = response.getBody().getJsonArray("accounts");
+      for (Object o : accountsJson) {
+        if (o instanceof JsonObject) {
+          JsonObject accountJson = (JsonObject) o;
+          final Item item = new Item().withItemId(accountJson.getString("itemId"));
+          final Charge charge = getCharge(accountJson, item);
+          amount += charge.getChargeAmount().getAmount().doubleValue();
+          if (includeCharges) {
+            charges.add(charge);
+          }
+        }
+      }
+    }
+
+    account.setTotalCharges(new TotalCharges()
+        .withAmount(amount)
+        .withIsoCurrencyCode("USD"));
+
+    return account;
+  }
+
+  private Charge getCharge(JsonObject chargeJson, Item item) {
+    return new Charge()
+        .withItem(item)
+        .withAccrualDate(new DateTime(chargeJson.getString("dateCreated")).toDate())
+        .withChargeAmount(new TotalCharges().withAmount(chargeJson.getDouble("remaining")).withIsoCurrencyCode("USD"))
+        .withState(chargeJson.getJsonObject("paymentStatus", new JsonObject().put("name",  "Unknown")).getString("name"))
+        .withFeeFineId(chargeJson.getString("feeFineId"));
+  }
+
+  private CompletableFuture<Account> lookupChargeItem(HttpClientInterface httpClient, Charge charge, Account account, Map<String, String> okapiHeaders) {
+    return getChargeItem(charge, httpClient, okapiHeaders)
+        .thenApply(this::verifyAndExtractBody)
+        .thenCompose(chargeItem ->
+          itemLookup(chargeItem, charge, account, httpClient, okapiHeaders)
+        );
+  }
+
+  private CompletableFuture<Account> itemLookup(JsonObject chargeItem, Charge charge,
+      Account account, HttpClientInterface httpClient,
+      Map<String, String> okapiHeaders) {
+    try {
+      final String barcode = chargeItem.getString("barcode");
+
+      return httpClient.request("/inventory/items?query=barcode%3D%3D" + barcode, okapiHeaders)
+          .thenCompose(invResponse -> {
+            if (Response.isSuccess(invResponse.getCode())) {
+              if (invResponse.getBody().getInteger("totalRecords").intValue() > 0) {
+                final JsonObject item = invResponse.getBody().getJsonArray("items").getJsonObject(0);
+                return getHoldingsRecord(item, httpClient, okapiHeaders)
+                  .thenApply(this::verifyAndExtractBody)
+                  .thenCompose(holdingsRecord -> getInstance(holdingsRecord, httpClient, okapiHeaders))
+                  .thenApply(this::verifyAndExtractBody)
+                  .thenApply(instance -> getItem(item, instance))
+                  .thenApply(itemObject -> updateChargeItem(charge, itemObject, account));
+              } else {
+                charge.getItem().withTitle(chargeItem.getString("instance"));
+                return CompletableFuture.completedFuture(account);
+              }
+            } else {
+              charge.getItem().withTitle(chargeItem.getString("instance"));
+              return CompletableFuture.completedFuture(account);
+            }
+          });
+    } catch (Exception e) {
+      throw new CompletionException(e);
+    }
+  }
+
+  private CompletableFuture<Response> getChargeItem(Charge charge,
+      HttpClientInterface httpClient, Map<String, String> okapiHeaders) {
+    try {
+      return httpClient.request("/chargeitem/" + charge.getItem().getItemId(), okapiHeaders);
+    } catch (Exception e) {
+      throw new CompletionException(e);
+    }
+  }
+
+  private CompletableFuture<Response> getHoldingsRecord(JsonObject item,
+      HttpClientInterface httpClient, Map<String, String> okapiHeaders) {
+    try {
+      return httpClient.request("/holdings-storage/holdings/" + item.getString("holdingsRecordId"), okapiHeaders);
+    } catch (Exception e) {
+      throw new CompletionException(e);
+    }
+  }
+
+  private CompletableFuture<Response> getInstance(JsonObject holdingsRecord,
+      HttpClientInterface httpClient, Map<String, String> okapiHeaders) {
+    try {
+      return httpClient.request("/inventory/instances/" + holdingsRecord.getString("instanceId"), okapiHeaders);
+    } catch (Exception e) {
+      throw new CompletionException(e);
+    }
+  }
+
+  private JsonObject verifyAndExtractBody(Response response) {
+    if (!Response.isSuccess(response.getCode())) {
+      throw new CompletionException(new HttpException(response.getCode(), response.getError().toString()));
+    }
+
+    return response.getBody();
+  }
+
+  private Item getItem(JsonObject item, JsonObject instance) {
+    final String itemId = item.getString("id");
+    final JsonObject composite = new JsonObject()
+        .put("contributors", instance.getJsonArray("contributors"))
+        .put("instanceId", instance.getString("id"))
+        .put("title", instance.getString("title"));
+
+    return getItem(itemId, composite);
+  }
+
+  private Account updateChargeItem(Charge charge, Item item, Account account) {
+    charge.withItem(item);
+
+    return account;
+  }
+
+  private CompletableFuture<Account> lookupFeeFine(HttpClientInterface httpClient,
+      Charge charge, Account account, Map<String, String> okapiHeaders) {
+    return getFeeFine(charge, httpClient, okapiHeaders)
+        .thenApply(this::verifyAndExtractBody)
+        .thenApply(feefine -> updateChargeWithFeeFine(feefine, charge, account));
+  }
+
+  private CompletableFuture<Response> getFeeFine(Charge charge,
+      HttpClientInterface httpClient, Map<String, String> okapiHeaders) {
+    try {
+      return httpClient.request("/feefines/" + charge.getFeeFineId(), okapiHeaders);
+    } catch (Exception e) {
+      throw new CompletionException(e);
+    }
+  }
+
+  private Account updateChargeWithFeeFine(JsonObject feeFine, Charge charge, Account account) {
+    charge.withReason(feeFine.getString("feeFineType"));
+
+    return account;
   }
 
   private int getLimit(boolean includeItem) {